file name changed
"invalid character in Bundle Identifier. This string must be a uniform type identifier (UTI) that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters."

The same error is displaying for the other project because there are space in the used app name. To prevent this error I changed the name of the project and other required file and path.